flight_task/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.0,<4.0
